More WIP attempting to wrap DAG for resource control
